+++ Changelog.md
@@ -1,6 +1,7 @@
 ### 0.5.5 (unreleased)
 
 Language Features:
+ * Add support for accessors for mappings with string or byte key types
  * Meta programming: Provide access to the name of contracts via ``type(C).name``.